requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
markdownify>=0.11.0
pyyaml>=6.0
tqdm>=4.66.0
pytest>=7.4.0
black>=23.12.0
flake8>=7.0.0
